fix: calculate uat map data using only cui
The uat spending is based on the cui. There can by entities in that uat that are not managed by the uat, but by the ministry or other entities. More precisely, don't use the location of the entity to group them and calculate the total spending of that uat. Use the uat from the entity, as the uat has also an entity with a cui.